codecraft/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-modeltranslation==0.19.12
django-multiselectfield==0.1.13
django-phonenumber-field==8.0.0
idna==3.10
phonenumbers==8.13.55
pillow==11.1.0
polib==1.2.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
whitenoise==6.9.0
